+++ src/infra/timescale_db/models/sensor2.py
@@ -28,14 +28,14 @@ class Sensor2(Base, CreateTimestampMixin):
     frequency_status_3: Mapped[int] = mapped_column(Integer, nullable=False, comment="Состояние ПЧ 3")
     frequency_status_4: Mapped[int] = mapped_column(Integer, nullable=False, comment="Состояние ПЧ 4")
 
-    frequency_torque_1: Mapped[int] = mapped_column(Integer, nullable=False, comment="Момент ПЧ 1")
-    frequency_torque_2: Mapped[int] = mapped_column(Integer, nullable=False, comment="Момент ПЧ 2")
-    frequency_torque_3: Mapped[int] = mapped_column(Integer, nullable=False, comment="Момент ПЧ 3")
-    frequency_torque_4: Mapped[int] = mapped_column(Integer, nullable=False, comment="Момент ПЧ 4")
+    frequency_torque_1: Mapped[float] = mapped_column(Float, nullable=False, comment="Момент ПЧ 1")
+    frequency_torque_2: Mapped[float] = mapped_column(Float, nullable=False, comment="Момент ПЧ 2")
+    frequency_torque_3: Mapped[float] = mapped_column(Float, nullable=False, comment="Момент ПЧ 3")
+    frequency_torque_4: Mapped[float] = mapped_column(Float, nullable=False, comment="Момент ПЧ 4")
 
-    converter_frequency_1: Mapped[int] = mapped_column(Integer, nullable=False, comment="Частота ПЧ 1")
-    converter_frequency_2: Mapped[int] = mapped_column(Integer, nullable=False, comment="Частота ПЧ 2")
-    converter_frequency_3: Mapped[int] = mapped_column(Integer, nullable=False, comment="Частота ПЧ 3")
-    converter_frequency_4: Mapped[int] = mapped_column(Integer, nullable=False, comment="Частота ПЧ 4")
+    converter_frequency_1: Mapped[float] = mapped_column(Float, nullable=False, comment="Частота ПЧ 1")
+    converter_frequency_2: Mapped[float] = mapped_column(Float, nullable=False, comment="Частота ПЧ 2")
+    converter_frequency_3: Mapped[float] = mapped_column(Float, nullable=False, comment="Частота ПЧ 3")
+    converter_frequency_4: Mapped[float] = mapped_column(Float, nullable=False, comment="Частота ПЧ 4")
 
     timestamp: Mapped[datetime] = mapped_column(nullable=False, comment="метка времени измерения")
